Handle OL3 measure pointer event independently. Fixes #218
Previously the logic to update the distance and position the tooltips
when measuring was included in the main pointer move event handler which
deals with highlighting and showing info for features. This was causing
issues as the measure logic was not always being called due to the hover
logic before it.